requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
python-barcode>=0.15.0
Pillow>=10.0.0
requests>=2.31.0
urllib3>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0